Create references.Rmd
The bibliography will render at the bottom of each page for items mentioned on the page, and once for all items in the “book” as the last item on the last page. Make a new Rmd with just a single heading ‘# References’ and if it is the last Rmd alphabetically it will contain the bibliography.